requirements.txt
schedule
smtplib
python-dotenv
nltk
